pyproject.toml
[project]
name = "pygame-new"
version = "0.1.0"
description = "an example of a pygame repo"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pygame-ce>=2.5.3",
]

[dependency-groups]
dev = [
    "pygbag>=0.9.2",
    "pytest>=8.4.0",
    "ruff>=0.11.13",
]

[tool.ruff.lint]
select = [
    "ANN",  # Require return type annotations for public functions
    "D", # Enable docstring rules
]

# Enable the docstring line length check
extend-select = ["W505"]

[tool.ruff.lint.per-file-ignores]
# Disable all docstring rules for tests.
# Tests should have pretty descriptive naming already. 
# See Unit Testing Principles, Practices, and Patterns (ISBN-13 978-1617296277)
"tests/*" = ["D"]

[tool.ruff.lint.pycodestyle]
# max line length for docstrings
max-doc-length = 88

# This is required to use an editable installation of the `pygame-new` library,
# which is required to use pytest in this configuration.
#
# See <https://docs.pytest.org/en/stable/explanation/goodpractices.html>.
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
